+++ front-end/src/styles/AddExpense.css
@@ -1,0 +1,19 @@
+form {
+    @apply flex flex-col space-y-3 mt-10 bg-cyan-50 rounded-2xl p-8;
+}
+
+form div {
+    @apply ml-6;
+}
+
+.formBtn {
+    @apply m-auto;
+}
+
+.formBtn button {
+    @apply w-24 h-8 rounded-full bg-cyan-100 hover:bg-orange-100 active:bg-orange-200 mt-2;
+}
+
+#splitOptions {
+    @apply fixed top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 bg-cyan-200 rounded-2xl p-8;
+}
